+++ CHANGELOG.md
@@ -152,6 +152,7 @@
 - Enh #153: Add `HasOffset` trait and corresponding tests for managing SVG `offset` attribute (@terabytesoftw)
 - Enh #154: Add `HasStopColor` trait and corresponding tests for managing SVG `stop-color` attribute (@terabytesoftw)
 - Enh #155: Add `HasStopOpacity` trait and corresponding tests for managing SVG `stop-opacity` attribute (@terabytesoftw)
+- Enh #156: Add `Stop` class and corresponding tests for SVG `<stop>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024